requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.5.0
idna==2.7
jira==2.0.0
oauthlib==2.1.0
pbr==5.0.0
pycparser==2.19
PyJWT==1.6.4
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.24
